requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
csscompressor==0.9.5
daiquiri==2.1.1
decorator==4.4.2
defusedxml==0.6.0
Django==1.11
django-bootstrap-datepicker-plus==3.0.5
django-datetimepicker==3.14
django-decorator-include==2.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
entrypoints==0.3
html5lib==1.0b10
httplib2==0.10.3
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-datatables==0.3.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-nbutils==0.1.3
jupyter-require==0.6.1
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
olefile==0.44
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==4.1.0
pisa==3.0.33
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.7.1
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
PyPDF2==1.26.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2017.2
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
reportlab==3.5.42
scipy==1.4.1
Send2Trash==1.5.0
six==1.10.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
virtualenv==16.1.0
wcwidth==0.1.9
webencodings==0.5.1
win-unicode-console==0.5
xhtml2pdf==0.2b1
zipp==1.2.0
